pyproject.toml
[project]
name = "compgene"
version = "0.1.0"
description = "Comparative genomics automated analysis pipeline"
requires-python = ">=3.11"
dependencies = [
    "snakemake>=9.14",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1.0"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["workflow/adapters", "tests"]
python_files = "test_*.py"
